JpegtoPDFConversion/requirements.txt
Pillow==10.3.0
reportlab==4.1.0
